requirements.txt
alembic
Flask
Flask_Bcrypt
Flask_Cors
Flask_Migrate
flask_sqlalchemy
ipdb
python-dotenv
SQLAlchemy
SQLAlchemy_serializer
gunicorn
psycopg2
